tests/wpt/web-platform-tests/css/css-contain/container-queries/container-units-shadow.html

<!doctype html>
<title>Container Relative Units: Shadow DOM</title>
<link rel="help" href="https://drafts.csswg.org/css-contain-3/#container-lengths">
<link rel="help" href="https://drafts.csswg.org/css-contain-3/#container-queries">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/declarative-shadow-dom-polyfill.js"></script>
<script src="support/cq-testcommon.js"></script>
<style>
  #outer {
    container-type: size;
    width: 200px;
    height: 200px;
  }
  #direct {
    container-type: inline-size;
    width: 50cqw;
    height: 50cqh;
  }
  #nondirect {
    width: 10cqw;
    height: 10cqh;
    background: green;
  }
</style>
<div id=outer>
  <div>
    <template shadowroot="open">
      <style>
        #inner {
          container-type: size;
          width: 30px;
          height: 30px;
        }
      </style>
      <div id=inner>
        <slot></slot>
      </div>
    </template>
    <div>
      <div id=direct>
        <div id=nondirect>
        </div>
      </div>
    </div>
  </div>
</div>
<script>
  setup(() => {
    assert_implements_container_queries();
    polyfill_declarative_shadow_dom(document);
  });

  test(() => {
    let cs = getComputedStyle(direct);
    assert_equals(cs.width, '100px');
    assert_equals(cs.height, '100px');
  }, 'Direct slotted child queries shadow-including ancestors');

  test(() => {
    let cs = getComputedStyle(nondirect);
    assert_equals(cs.width, '10px'); // #direct
    assert_equals(cs.height, '20px'); // #outer
  }, 'Nondirect slotted child queries shadow-including ancestors');
</script>
